+++ hc/api/tests/test_delete_check.py
@@ -6,11 +6,10 @@ class DeleteCheckTestCase(BaseTestCase):
     def setUp(self):
         super().setUp()
         self.check = Check.objects.create(project=self.project)
+        self.url = f"/api/v1/checks/{self.check.code}"
 
     def test_it_works(self):
-        r = self.client.delete(
-            "/api/v1/checks/%s" % self.check.code, HTTP_X_API_KEY="X" * 32
-        )
+        r = self.client.delete(self.url, HTTP_X_API_KEY="X" * 32)
         self.assertEqual(r.status_code, 200)
         self.assertEqual(r["Access-Control-Allow-Origin"], "*")
 
@@ -23,6 +22,10 @@ class DeleteCheckTestCase(BaseTestCase):
         self.assertEqual(r.status_code, 404)
 
     def test_it_handles_options(self):
-        r = self.client.options("/api/v1/checks/%s" % self.check.code)
+        r = self.client.options(self.url)
         self.assertEqual(r.status_code, 204)
         self.assertIn("DELETE", r["Access-Control-Allow-Methods"])
+
+    def test_it_handles_missing_api_key(self):
+        r = self.client.delete(self.url)
+        self.assertContains(r, "missing api key", status_code=401)